Rename pointers.js module to gestures.js
